[frontend] useBuildAttributesOutcome
